Mega refactor! 🎉
The underlying concepts are the same, but this refactor creates a new timer for each class we're looking to amend.

In addition to this, we had to create a "total" counter to stop the polling for an element after 10 second if it hasn't appeared on the page. This happens with things like the Breadcrumbs and Dropdown Filter when we're not on a Product or Category page.